Separate extensions into different files
